Update_o_final?

+++ SMSNinjaUI/SNCallActionViewController.m
@@ -11,10 +11,10 @@
 {
 	[phoneAction release];
 	phoneAction = nil;
-    
+
 	[flag release];
 	flag = nil;
-    
+
 	[super dealloc];
 }
 
@@ -43,41 +43,41 @@
 {
 	UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:@"any-cell"];
 	if (cell == nil) cell = [[[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:@"any-cell"] autorelease];
-    for (UIView *subview in [cell.contentView subviews])
-        [subview removeFromSuperview];
-    cell.textLabel.text = nil;
-    cell.accessoryView = nil;
-    cell.accessoryType = UITableViewCellAccessoryNone;
-    
-    switch (indexPath.row)
-    {
-        case 0:
-            cell.textLabel.text = NSLocalizedString(@"Disconnect", @"Disconnect");
-            if ([self.phoneAction isEqualToString:@"1"]) cell.accessoryType = UITableViewCellAccessoryCheckmark;
-            else cell.accessoryType = UITableViewCellAccessoryNone;
-            break;
-        case 1:
-            cell.textLabel.text = NSLocalizedString(@"Ignore", @"Ignore");
-            if ([self.phoneAction isEqualToString:@"2"]) cell.accessoryType = UITableViewCellAccessoryCheckmark;
-            else cell.accessoryType = UITableViewCellAccessoryNone;
-            break;
-        case 2:
-            cell.textLabel.text = NSLocalizedString(@"Let go", @"Let go");
-            if ([self.phoneAction isEqualToString:@"3"]) cell.accessoryType = UITableViewCellAccessoryCheckmark;
-            else cell.accessoryType = UITableViewCellAccessoryNone;
-            break;
-    }
+	for (UIView *subview in [cell.contentView subviews])
+		[subview removeFromSuperview];
+	cell.textLabel.text = nil;
+	cell.accessoryView = nil;
+	cell.accessoryType = UITableViewCellAccessoryNone;
+
+	switch (indexPath.row)
+	{
+		case 0:
+			cell.textLabel.text = NSLocalizedString(@"Disconnect", @"Disconnect");
+			if ([self.phoneAction isEqualToString:@"1"]) cell.accessoryType = UITableViewCellAccessoryCheckmark;
+			else cell.accessoryType = UITableViewCellAccessoryNone;
+			break;
+		case 1:
+			cell.textLabel.text = NSLocalizedString(@"Ignore", @"Ignore");
+			if ([self.phoneAction isEqualToString:@"2"]) cell.accessoryType = UITableViewCellAccessoryCheckmark;
+			else cell.accessoryType = UITableViewCellAccessoryNone;
+			break;
+		case 2:
+			cell.textLabel.text = NSLocalizedString(@"Let go", @"Let go");
+			if ([self.phoneAction isEqualToString:@"3"]) cell.accessoryType = UITableViewCellAccessoryCheckmark;
+			else cell.accessoryType = UITableViewCellAccessoryNone;
+			break;
+	}
 	return cell;
 }
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
 	[tableView deselectRowAtIndexPath:indexPath animated:YES];
-    
+
 	if ([tableView cellForRowAtIndexPath:indexPath].accessoryType == UITableViewCellAccessoryCheckmark)
 	{
 		[tableView cellForRowAtIndexPath:indexPath].accessoryType = UITableViewCellAccessoryNone;
-        
+
 		self.phoneAction = nil;
 		self.phoneAction = @"0";
 	}
@@ -85,9 +85,9 @@
 	{
 		[tableView cellForRowAtIndexPath:indexPath].accessoryType = UITableViewCellAccessoryCheckmark;
 		[tableView cellForRowAtIndexPath:[NSIndexPath indexPathForRow:(1 - indexPath.row) inSection:0]].accessoryType = UITableViewCellAccessoryNone;
-        
+
 		self.phoneAction = nil;
-		self.phoneAction = [NSString stringWithFormat:@"%d", indexPath.row + 1];
+		self.phoneAction = [NSString stringWithFormat:@"%ld", (long)(indexPath.row + 1)];
 	}
 	else if ([self.flag isEqualToString:@"private"] && [tableView cellForRowAtIndexPath:indexPath].accessoryType == UITableViewCellAccessoryNone)
 	{
@@ -97,19 +97,19 @@
 			if (i != indexPath.row)
 				[tableView cellForRowAtIndexPath:[NSIndexPath indexPathForRow:i inSection:0]].accessoryType = UITableViewCellAccessoryNone;
 		}
-        
+
 		self.phoneAction = nil;
-		self.phoneAction = [NSString stringWithFormat:@"%d", indexPath.row + 1];
+		self.phoneAction = [NSString stringWithFormat:@"%ld", (long)(indexPath.row + 1)];
 	}
-    
-    id viewController = [self.navigationController.viewControllers objectAtIndex:([self.navigationController.viewControllers count] - 2)];
-    [viewController setPhoneAction:self.phoneAction];
-    if ([viewController isKindOfClass:[SNTimeViewController class]]) [((SNTimeViewController *)viewController)->settingsTableView reloadData];
-    else [((UITableViewController *)viewController).tableView reloadData];
+
+	id viewController = [self.navigationController.viewControllers objectAtIndex:([self.navigationController.viewControllers count] - 2)];
+	[viewController setPhoneAction:self.phoneAction];
+	if ([viewController isKindOfClass:[SNTimeViewController class]]) [((SNTimeViewController *)viewController)->settingsTableView reloadData];
+	else [((UITableViewController *)viewController).tableView reloadData];
 }
 
 - (NSString *)tableView:(UITableView *)tableView titleForFooterInSection:(NSInteger)section
 {
-    return NSLocalizedString(@"Uncheck to disable", @"Uncheck to disable");
+	return NSLocalizedString(@"Uncheck to disable", @"Uncheck to disable");
 }
 @end
